Fix bugs missed in the code review

+++ src/net/serialization.cc
@@ -30,12 +30,12 @@ uint16_t get_uint16(const char * data)
   return be16toh(*reinterpret_cast<const uint16_t *>(data));
 }
 
-uint16_t get_uint32(const char * data)
+uint32_t get_uint32(const char * data)
 {
   return be32toh(*reinterpret_cast<const uint32_t *>(data));
 }
 
-uint16_t get_uint64(const char * data)
+uint64_t get_uint64(const char * data)
 {
   return be64toh(*reinterpret_cast<const uint64_t *>(data));
 }
